Initial commit.
Files SymbolicEquationsOfMotion.m contain the code for generating equations of motion for both dynamic and static cases.  It also writes and reads a table called StaticVarTable.csv which you can fill in values for the parameters of mass, force, length, moments etc.  Leave as "NaN" if this is a variable that will be solved for..

The file InverseKinematicsOpti.m is a crude optimization for getting the pose of the leg (knee, hip and ankle angles).